requirements.txt
openai-whisper==20231117
bitsandbytes==0.42.0
datasets==2.16.1
torchaudio==2.1.2
librosa==0.10.1
soundfile==0.12.1; platform_system=="Windows"
tiktoken @ git+https://github.com/swiss-german-speech-to-text/tiktoken@main
transformers==4.37.2